move Mortality_Polynomial to Impact, rename to PolynomialImpact, clean up formatting
